Fix drag handle tap targets (especially on mobile)
There was a bug where sticky positioned <th> elements were higher than
the overlapping row/column resize drag handle in the z axis. This
occurred even though the drag handles had absolute positioning, and
higher z-index than the <th> elements. But because the drag handles are
children of the <th> elements, they could not be higher than the sibling
of their ancestor in a different stacking context.

The result was that only half of the drag handle was clickable, which
made the tap targets much too small. This change fixes that bug.

Thanks to @lsnow99 who saved me from going down a bad route with
hard-coded, decreasing z indices for <th> elements.